consolidated duplicate postal code formatting code
