+++ frontend/src/features/mapSideBar/components/tabs/ParcelDetails.tsx
@@ -13,7 +13,7 @@ import React, { CSSProperties, Dispatch, SetStateAction, useState } from 'react'
 import { FaEdit } from 'react-icons/fa';
 import { useAppSelector } from 'store';
 import styled from 'styled-components';
-import { mapLookupCode } from 'utils';
+import { mapLookupCode, postalCodeFormatter } from 'utils';
 
 import { tabStyles } from './TabStyles';
 
@@ -53,19 +53,6 @@ export const ParcelDetails: React.FC<any> = (props: IParcelDetailsProps) => {
   const administrativeAreas = _.filter(lookupCodes, (lookupCode: ILookupCode) => {
     return lookupCode.type === API.AMINISTRATIVE_AREA_CODE_SET_NAME;
   }).map(mapLookupCode);
-
-  /**
-   * postalCodeFormatter takes the specified postal code and formats it with a space in the middle
-   * @param {string} postal The target postal to be formatted
-   */
-  const postalCodeFormatter = (postal: string) => {
-    const regex = /([a-zA-z][0-9][a-zA-z])[\s-]?([0-9][a-zA-z][0-9])/;
-    const format = postal.match(regex);
-    if (format !== null && format.length === 3) {
-      postal = `${format[1]} ${format[2]}`;
-    }
-    return postal.toUpperCase();
-  };
 
   return (
     <div className="identification">
