+++ frontend/src/utils/index.ts
@@ -2,5 +2,6 @@ export * from './formatAddress';
 export * from './fromQueryString';
 export * from './mapLookupCode';
 export * from './numberFormatUtils';
+export * from './postalCodeFormatter';
 export * from './toQueryString';
 export * from './utils';
